move action helper to utils, fix buttons
